+++ app/exceptions/base_exception.py
@@ -9,6 +9,11 @@ class AlreadyExistsError(Exception):
         super().__init__(self.message)
 
 class ExternalServiceError(Exception):
-    def __init__(self, service:str, detail: str):
+    def __init__(self, service: str, detail: str):
         self.message = f"{service} error: {detail}"
+        super().__init__(self.message)
+
+class ForbiddenError(Exception):
+    def __init__(self, entity: str, identifier: str | int, user_id: int):
+        self.message = f"{entity} with id '{identifier}' cannot be modified by user '{user_id}'"
         super().__init__(self.message)